migrated notation tests into new file
